Further refactor of directories to better encapsulate responsibilities of different components, also README update to explain this
