+++ src/features/CourtCaseDetails/CourtCaseDetails.stories.tsx
@@ -3,8 +3,8 @@ import { expect } from "@storybook/jest"
 import { ComponentMeta, ComponentStory } from "@storybook/react"
 import { within } from "@storybook/testing-library"
 import { format } from "date-fns"
-import CourtCase from "entities/CourtCase"
-import Trigger from "entities/Trigger"
+import CourtCase from "services/entities/CourtCase"
+import Trigger from "services/entities/Trigger"
 import CourtCaseAho from "../../../test/testFixtures/database/data/error_list_aho.json"
 import CourtCaseDetails from "./CourtCaseDetails"
