+++ src/middleware/withAuthentication/withAuthentication.ts
@@ -1,6 +1,6 @@
-import User from "../../entities/User"
-import getDataSource from "lib/getDataSource"
-import parseJwtCookie from "lib/parseJwtCookie"
+import User from "../../services/entities/User"
+import getDataSource from "services/getDataSource"
+import parseJwtCookie from "middleware/withAuthentication/parseJwtCookie"
 import { GetServerSideProps, GetServerSidePropsContext, GetServerSidePropsResult } from "next"
 import { ParsedUrlQuery } from "querystring"
 import AuthenticationServerSidePropsContext from "types/AuthenticationServerSidePropsContext"
