+++ src/pages/court-cases/[courtCaseId]/index.tsx
@@ -1,7 +1,7 @@
 import Layout from "components/Layout"
-import CourtCase from "entities/CourtCase"
-import User from "entities/User"
-import getDataSource from "lib/getDataSource"
+import CourtCase from "services/entities/CourtCase"
+import User from "services/entities/User"
+import getDataSource from "services/getDataSource"
 import { withAuthentication, withMultipleServerSideProps } from "middleware"
 import type { GetServerSidePropsContext, GetServerSidePropsResult, NextPage } from "next"
 import Head from "next/head"
