+++ src/services/entities/BaseEntity.ts
