+++ src/services/getDataSource.ts
@@ -1,9 +1,9 @@
-import CourtCase from "../entities/CourtCase"
+import CourtCase from "./entities/CourtCase"
 import { DataSource, DataSourceOptions } from "typeorm"
 import DatabaseConfig from "./DatabaseConfig"
-import User from "../entities/User"
-import Trigger from "../entities/Trigger"
-import Note from "../entities/Note"
+import User from "./entities/User"
+import Trigger from "./entities/Trigger"
+import Note from "./entities/Note"
 
 const databaseConfig: DatabaseConfig = {
   host: process.env.DB_HOST ?? process.env.DB_AUTH_HOST ?? "localhost",
