+++ test/services/getCourtCase.integration.test.ts
@@ -1,8 +1,8 @@
 import { expect } from "@jest/globals"
 import parseAhoXml from "@moj-bichard7-developers/bichard7-next-core/build/src/parse/parseAhoXml/parseAhoXml"
 import { DataSource } from "typeorm"
-import CourtCase from "../../src/entities/CourtCase"
-import getDataSource from "../../src/lib/getDataSource"
+import CourtCase from "../../src/services/entities/CourtCase"
+import getDataSource from "../../src/services/getDataSource"
 import getCourtCase from "../../src/services/getCourtCase"
 import { isError } from "../../src/types/Result"
 import CourtCaseCase from "../testFixtures/database/data/error_list.json"
