+++ test/services/listCourtCases.integration.test.ts
@@ -13,8 +13,8 @@ import {
 } from "../testFixtures/database/insertCourtCases"
 import insertException from "../testFixtures/database/manageExceptions"
 import { isError } from "../../src/types/Result"
-import CourtCase from "../../src/entities/CourtCase"
-import getDataSource from "../../src/lib/getDataSource"
+import CourtCase from "../../src/services/entities/CourtCase"
+import getDataSource from "../../src/services/getDataSource"
 import { insertTriggers, TestTrigger } from "../testFixtures/database/manageTriggers"
 
 jest.setTimeout(100000)
